refactor: move OakLink to SharedComponents
