+++ src/components/AppComponents/AppHeader/AppHeader.tsx
@@ -2,11 +2,11 @@ import { FC, useRef } from "react";
 
 import Logo from "@/components/AppComponents/Logo";
 import { HeaderProps } from "@/components/AppComponents/Layout/Layout";
-import OakLink from "@/components/OakLink";
+import OakLink from "@/components/SharedComponents/OakLink";
 import { AppHeaderMenu } from "@/components/AppComponents/AppHeaderMenu";
 import { useMenuContext } from "@/context/Menu";
 import AppHeaderBurgerMenuSections from "@/components/AppComponents/AppHeaderBurgerMenuSections";
-import { ActiveLinkUnderline } from "@/components/OakLink/OakLink";
+import { ActiveLinkUnderline } from "@/components/SharedComponents/OakLink/OakLink";
 import IconButton from "@/components/SharedComponents/Button/IconButton";
 import Flex from "@/components/SharedComponents/Flex";
 import Box from "@/components/SharedComponents/Box";
