+++ src/components/GenericPagesComponents/WebinarRegistration/WebinarRegistration.tsx
@@ -3,7 +3,7 @@ import { FC, useId } from "react";
 import NewsletterForm, {
   useNewsletterForm,
 } from "@/components/GenericPagesComponents/NewsletterForm";
-import OakLink from "@/components/OakLink";
+import OakLink from "@/components/SharedComponents/OakLink";
 import BoxBorders from "@/components/SharedComponents/SpriteSheet/BrushSvgs/BoxBorders";
 import { HeadingTag, P } from "@/components/SharedComponents/Typography";
 import CardTitle from "@/components/SharedComponents/Card/CardComponents/CardTitle";
