+++ src/components/SharedComponents/Breadcrumbs/Breadcrumbs.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 import styled from "styled-components";
 
 import { BreadcrumbJsonLd } from "@/browser-lib/seo/getJsonLd";
-import OakLink from "@/components/OakLink";
+import OakLink from "@/components/SharedComponents/OakLink";
 import { MaybeOakHref, ResolveOakHrefProps } from "@/common-lib/urls";
 import Icon from "@/components/SharedComponents/Icon";
 import UL from "@/components/SharedComponents/Typography/UL";
