+++ src/components/SharedComponents/PortableText/PortableText.tsx
@@ -14,7 +14,7 @@ import { PTActionTrigger } from "./PTActionTrigger";
 import errorReporter from "@/common-lib/error-reporter";
 import { resolveInternalHref } from "@/utils/portableText/resolveInternalHref";
 import { CTAInternalLinkEntry } from "@/common-lib/cms-types";
-import OakLink from "@/components/OakLink";
+import OakLink from "@/components/SharedComponents/OakLink";
 import getProxiedSanityAssetUrl from "@/common-lib/urls/getProxiedSanityAssetUrl";
 import { LI, OL, P, Span } from "@/components/SharedComponents/Typography";
 import AnchorTarget from "@/components/SharedComponents/AnchorTarget";
