+++ src/components/SharedComponents/SpriteSheet/BrushSvgs/BrushBorders/BrushBorders.tsx
@@ -5,7 +5,7 @@ import getColorByName from "@/styles/themeHelpers/getColorByName";
 import { OakColorName } from "@/styles/theme/types";
 import { getBreakpoint } from "@/styles/utils/responsive";
 import Svg from "@/components/SharedComponents/Svg";
-import { $selectedStyle } from "@/components/OakLink/OakLink";
+import { $selectedStyle } from "@/components/SharedComponents/OakLink/OakLink";
 
 const brushBorder = css<{ color: OakColorName }>`
   position: absolute;
