+++ src/components/TeacherComponents/CopyrightNotice/CopyrightNotice.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 import styled from "styled-components";
 
 import P from "@/components/SharedComponents/Typography/P";
-import OakLink from "@/components/OakLink";
+import OakLink from "@/components/SharedComponents/OakLink";
 import Box from "@/components/SharedComponents/Box";
 import { FontProps } from "@/styles/utils/typography";
 import Icon from "@/components/SharedComponents/Icon";
